Clean: W以上の違反を修正。

+++ spec/rails_helper.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 ENV['RAILS_ENV'] = 'test'
 require File.expand_path('../config/environment', __dir__)
 # 本番環境でDBのテーブルから全ての行を削除するのを防ぐ
-abort("The Rails environment is running in production mode!") if Rails.env.production?
+abort('The Rails environment is running in production mode!') if Rails.env.production?
 require 'rspec/rails'
 
 # support/config配下(テスト用のヘルパーメソッドを使う際に記述するファイル置き場)のファイルを読み込み
@@ -56,9 +56,6 @@ RSpec.configure do |config|
 
   # テスト終了時に画像を削除する
   config.after(:all) do
-    if Rails.env.test?
-      FileUtils.rm_rf(Dir["#{Rails.root}/public/uploads_#{Rails.env}/"])
-    end
+    FileUtils.rm_rf(Dir["#{Rails.root}/public/uploads_#{Rails.env}/"]) if Rails.env.test?
   end
-
 end
